my_fastapi_app/requirements.txt
chardet==5.1.0
faiss_cpu==1.7.3
fastapi==0.92.0
numpy==1.22.4
opencv_python==4.7.0.72
pandas==1.5.3
Pillow==9.4.0
pydantic==1.9.1
torch==1.13.1
torchvision==0.14.1
uvicorn==0.17.6
python-multipart
